+++ src/ram_lib/managers/getters/__init__.py
@@ -1,2 +1,3 @@
 """Exposed getters classes and methods."""
 from ram_lib.managers.getters.ram_getter import RamGetter
+from ram_lib.managers.getters.use_getter import UseGetter