+++ tests/mock_ram_lib/__init__.py
@@ -1,4 +1,4 @@
 """Exposed mock_ram_lib classes and methods."""
-from mock_ram_lib.mock_ram_manager import mock_ram_memories, mock_ram_memories_entities
+from mock_ram_lib.mock_ram_manager import mock_ram_memories, mock_ram_memories_entities, use_ram, use_ram_formatted
 
 mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"